addAsset now using asset controller
